Closes #5, enhancements and small fixes. Related problem with disp_clear fixed. Action update.
