+++ .github/workflows/gh-pages.yml
@@ -1,8 +1,8 @@
 name: GitHub Pages
 on:
-  push:
-    branches:
-      - main
+  release:
+    types: [published]
+
 jobs:
   build-and-deploy:
     runs-on: ubuntu-latest
