+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "chip.ts",
   "description": "A CHIP-8 emulator in TypeScript",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "private": true,
   "repository": {
     "type": "git",
